1. File 완료 2. Vimeo 저장, 조회 확인해야함 3. Member Idx Type Long -> Int 변경

+++ src/main/java/softfocus/space/conference/module/today/repository/TodayRepository.java
@@ -8,5 +8,5 @@ import java.time.LocalDate;
 import java.util.Optional;
 
 public interface TodayRepository extends JpaRepository<Today, Long>, QuerydslPredicateExecutor<Today> {
-    Optional<Today> findByDayAndMember_Idx(LocalDate day, Long member_idx);
+    Optional<Today> findByDayAndMember_Idx(LocalDate day, Integer member_idx);
 }
